Another set of changes to address new compiler warnings where size_t bigger than long. Expect more. 	modified:   src/bin/dwarfdump/dd_command_options.c 	modified:   src/bin/dwarfdump/dd_esb.c 	modified:   src/bin/dwarfdump/dd_getopt.c 	modified:   src/bin/dwarfdump/dwarfdump.c 	modified:   src/bin/dwarfdump/print_macro.c 	modified:   src/lib/libdwarf/dwarf_debuglink.c 	modified:   src/lib/libdwarf/dwarf_macro5.c 	modified:   src/lib/libdwarf/dwarf_object_detector.c 	modified:   src/lib/libdwarf/dwarf_string.c 	modified:   src/lib/libdwarf/dwarf_string.h 	modified:   src/lib/libdwarf/dwarf_tsearchhash.c 	modified:   test/test_errmsglist.c

+++ src/lib/libdwarf/dwarf_macro5.c
@@ -219,7 +219,7 @@ _dwarf_skim_forms(Dwarf_Debug dbg,
             }
             v = strlen((char *) mdata) + 1;
             totallen += v;
-            mdata += v;
+            mdata    += v;
             }
             break;
         case DW_FORM_block:
@@ -706,8 +706,8 @@ static const char *
 construct_from_dir_and_name(const char *dir,
     const char *name)
 {
-    int truelen = 0;
-    char *final = 0;
+    size_t truelen = 0;
+    char  *final   = 0;
 
     /* Allow for NUL char and added /  */
     truelen = strlen(dir) + strlen(name) + 1 +1;
